refactor: Move procs into run
